got rid of errors and created seperate files for the classes in the framework.java file
