+++ src/cs3318/raytracing/Driver.java
@@ -45,9 +45,9 @@ class Driver  {
         // Parse the scene file
         String filename = dataFile != null ? dataFile : "defaultScene.txt";
 
-        InputStream is = null;
+        InputStream is;
         try {
-            is = new FileInputStream(new File(filename));
+            is = new FileInputStream(filename);
             ReadInput(is);
             is.close();
         } catch (IOException e) {
@@ -91,61 +91,70 @@ class Driver  {
 	          default:
 		        break scan;
 	          case StreamTokenizer.TT_WORD:
-	            if (st.sval.equals("sphere")) {
-                    Vector3D v = new Vector3D((float) getNumber(st), (float) getNumber(st), (float) getNumber(st));
-		            float r = (float) getNumber(st);
-		            objectList.add(new Sphere(currentSurface, v, r));
-			    } else
-			    if (st.sval.equals("eye")) {
-		            eye = new Vector3D((float) getNumber(st), (float) getNumber(st), (float) getNumber(st));
-			    } else
-			    if (st.sval.equals("lookat")) {
-		            lookat = new Vector3D((float) getNumber(st), (float) getNumber(st), (float) getNumber(st));
-			    } else
-			    if (st.sval.equals("up")) {
-		            up = new Vector3D((float) getNumber(st), (float) getNumber(st), (float) getNumber(st));
-			    } else
-			    if (st.sval.equals("fov")) {
-                    fov = (float) getNumber(st);
-			    } else
-			    if (st.sval.equals("background")) {
-                    background = Color.rgb((int) getNumber(st), (int) getNumber(st), (int) getNumber(st));
-			    } else
-			    if (st.sval.equals("light")) {
-			        float r = (float) getNumber(st);
-			        float g = (float) getNumber(st);
-			        float b = (float) getNumber(st);
-		            if (st.nextToken() != StreamTokenizer.TT_WORD) {
-                        throw new IOException(st.toString());
-                    }
-		            if (st.sval.equals("ambient")) {
-		                lightList.add(new Light(Light.AMBIENT, null, r, g, b));
-		            } else
-		            if (st.sval.equals("directional")) {
-		                Vector3D v = new Vector3D((float) getNumber(st), (float) getNumber(st), (float) getNumber(st));
-		                lightList.add(new Light(Light.DIRECTIONAL, v, r, g, b));
-		            } else
-		            if (st.sval.equals("point")) {
-		                Vector3D v = new Vector3D((float) getNumber(st), (float) getNumber(st), (float) getNumber(st));
-		                lightList.add(new Light(Light.POINT, v, r, g, b));
-		            } else {
-		                System.err.println("ERROR: in line "+st.lineno()+" at "+st.sval);
-		                throw new IOException(st.toString());
-		            }
-			    } else
-			    if (st.sval.equals("surface")) {
-			        float r = (float) getNumber(st);
-			        float g = (float) getNumber(st);
-			        float b = (float) getNumber(st);
-		            float ka = (float) getNumber(st);
-		            float kd = (float) getNumber(st);
-		            float ks = (float) getNumber(st);
-		            float ns = (float) getNumber(st);
-		            float kr = (float) getNumber(st);
-		            float kt = (float) getNumber(st);
-		            float index = (float) getNumber(st);
-		            currentSurface = new Surface(r, g, b, ka, kd, ks, ns, kr, kt, index);
-			    }
+				  switch (st.sval) {
+					  case "sphere": {
+						  Vector3D v = new Vector3D((float) getNumber(st), (float) getNumber(st), (float) getNumber(st));
+						  float r = (float) getNumber(st);
+						  objectList.add(new Sphere(currentSurface, v, r));
+						  break;
+					  }
+					  case "eye":
+						  eye = new Vector3D((float) getNumber(st), (float) getNumber(st), (float) getNumber(st));
+						  break;
+					  case "lookat":
+						  lookat = new Vector3D((float) getNumber(st), (float) getNumber(st), (float) getNumber(st));
+						  break;
+					  case "up":
+						  up = new Vector3D((float) getNumber(st), (float) getNumber(st), (float) getNumber(st));
+						  break;
+					  case "fov":
+						  fov = (float) getNumber(st);
+						  break;
+					  case "background":
+						  background = Color.rgb((int) getNumber(st), (int) getNumber(st), (int) getNumber(st));
+						  break;
+					  case "light": {
+						  float r = (float) getNumber(st);
+						  float g = (float) getNumber(st);
+						  float b = (float) getNumber(st);
+						  if (st.nextToken() != StreamTokenizer.TT_WORD) {
+							  throw new IOException(st.toString());
+						  }
+						  switch (st.sval) {
+							  case "ambient":
+								  lightList.add(new Light(Light.AMBIENT, null, r, g, b));
+								  break;
+							  case "directional": {
+								  Vector3D v = new Vector3D((float) getNumber(st), (float) getNumber(st), (float) getNumber(st));
+								  lightList.add(new Light(Light.DIRECTIONAL, v, r, g, b));
+								  break;
+							  }
+							  case "point": {
+								  Vector3D v = new Vector3D((float) getNumber(st), (float) getNumber(st), (float) getNumber(st));
+								  lightList.add(new Light(Light.POINT, v, r, g, b));
+								  break;
+							  }
+							  default:
+								  System.err.println("ERROR: in line " + st.lineno() + " at " + st.sval);
+								  throw new IOException(st.toString());
+						  }
+						  break;
+					  }
+					  case "surface": {
+						  float r = (float) getNumber(st);
+						  float g = (float) getNumber(st);
+						  float b = (float) getNumber(st);
+						  float ka = (float) getNumber(st);
+						  float kd = (float) getNumber(st);
+						  float ks = (float) getNumber(st);
+						  float ns = (float) getNumber(st);
+						  float kr = (float) getNumber(st);
+						  float kt = (float) getNumber(st);
+						  float index = (float) getNumber(st);
+						  currentSurface = new Surface(r, g, b, ka, kd, ks, ns, kr, kt, index);
+						  break;
+					  }
+				  }
 			    break;
 	        }
 	    }
